+++ astropy/coordinates/angle_utilities.py
@@ -26,21 +26,8 @@ import numpy as np
 from .errors import (IllegalHourWarning, IllegalHourError,
                      IllegalMinuteWarning, IllegalMinuteError,
                      IllegalSecondWarning, IllegalSecondError)
-from astropy.utils import format_exception
+from astropy.utils import format_exception, parsing
 from astropy import units as u
-
-TAB_HEADER = """# -*- coding: utf-8 -*-
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-
-# This file was automatically generated from ply. To re-generate this file,
-# remove it from this folder, then build astropy and run the tests in-place:
-#
-#   python setup.py build_ext --inplace
-#   pytest astropy/coordinates
-#
-# You can then commit the changes to this file.
-
-"""
 
 
 class _AngleParser:
@@ -164,15 +151,7 @@ class _AngleParser:
             raise ValueError(
                 f"Invalid character at col {t.lexpos}")
 
-        lexer_exists = os.path.exists(os.path.join(os.path.dirname(__file__),
-                                                   'angle_lextab.py'))
-
-        # Build the lexer
-        lexer = lex.lex(optimize=True, lextab='angle_lextab',
-                        outputdir=os.path.dirname(__file__))
-
-        if not lexer_exists:
-            cls._add_tab_header('angle_lextab')
+        lexer = parsing.lex(lextab='angle_lextab', package='astropy/coordinates')
 
         def p_angle(p):
             '''
@@ -314,29 +293,9 @@ class _AngleParser:
         def p_error(p):
             raise ValueError
 
-        parser_exists = os.path.exists(os.path.join(os.path.dirname(__file__),
-                                                    'angle_parsetab.py'))
-
-        parser = yacc.yacc(debug=False, tabmodule='angle_parsetab',
-                           outputdir=os.path.dirname(__file__),
-                           optimize=True, write_tables=True)
-
-        if not parser_exists:
-            cls._add_tab_header('angle_parsetab')
+        parser = parsing.yacc(tabmodule='angle_parsetab', package='astropy/coordinates')
 
         return parser, lexer
-
-    @classmethod
-    def _add_tab_header(cls, name):
-
-        lextab_file = os.path.join(os.path.dirname(__file__), name + '.py')
-
-        with open(lextab_file, 'r') as f:
-            contents = f.read()
-
-        with open(lextab_file, 'w') as f:
-            f.write(TAB_HEADER)
-            f.write(contents)
 
     def parse(self, angle, unit, debug=False):
         try: